Fix half-opened bottom sheet when returning from background.
The home-gesture on iOS devices without a physical home button can collide with the
DragGesture that is attached to the bottom sheet. The gesture allows to hide the
presented bottom sheet by dragging the top bar down to the edge of the screen.

However, when swiping up from the bottom to send the app to the background, the
gesture still receives changes. When the bottom sheet is hidden, this can lead to a
miscalculation of the sheet position. See the attached video to the PR.

The fix is to ignore the swipe-up gesture when the bottom sheet is hidden.